Don't import all cats instances in CornichonFeature
As there is a bunch of implicit in `CornichonFeature`'s scope, the
implicit resolution makes compilation slow and IDE user experience
really painfull.

To overcome this problem, this commit proposes to let the user the
choice of importing all cats instances by mixin `AllInstances` trait in
`CornichonFeature` or picking only the instances he needs if he is a
more advance user.