Fix date/date-time/time matchers thread safety
The "any-time", "any-date" & "any-date-time" matchers used the
`SimpleDateFormat` underneath, which is non-thread-safe for parsing due
to internal mutability. This caused errors in some cases when running
cornichon specs in a parallel way. E.g. the IntelliJ IDEA runner seems
to always run the specs serially, but SBT may run them in parallel,
causing the same set of specs to either succeed or fail depending on the
running environment.

Until Java 8 the common fix was to use a drop-in
replacement from Apache commons, `FastDateFormat`, implementing the same
`DateFormat` interface. But Java 8's java.time brought the newer,
immutable `DateTimeFormatter` class with default implementations for
most of the common temporal formats.

The added benefit is the less rigid checking against optional string
temporal value representation components, such as milliseconds in a
timestamp.