feat(core): sttp based client implementation
this replaces the current http4s client
with an implementation based on sttp.
the default backend for sttp is still using
http4s and the client is created in the same way,
so this should be backwards-compatible in terms of
the out-of-the-box client behaviour.
additionally, the netty async client is added
(https://sttp.softwaremill.com/en/latest/backends/monix.html#using-async-http-client)
as an alternative client option
(somewhat related to https://github.com/agourlay/cornichon/issues/391)